requirements.txt
cryptography>=41.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
jinja2>=3.1.0
pytest>=7.0.0

# Environment configuration
python-dotenv>=1.0.0

# Rate limiting for FastAPI
slowapi>=0.1.8

# Testing enhancements
pytest-xdist>=3.0.0       # Parallel test execution
pytest-cov>=4.0.0         # Test coverage reporting
pytest-asyncio>=0.21.0    # Async test support
freezegun>=1.2.0          # Datetime mocking for tests
